Proxy engine
This engine acts as a proxy and goes through a list of configured
engines, asking each of them if they are configured to handle the
current image. If they are, they handle the request. The list of engines
is treated as ordered. Engines listed first have a higher priority, since
they're asked first if they can handle the current image.

Bug: T119185
Change-Id: I774ec8353acbab12e8eeedb5e3c4caa3723086ba